chore: Break out FunctionSpy
