+++ tests/test_test_writer.py
@@ -3,15 +3,13 @@ from contextlib import ExitStack
 from pathlib import Path
 from unittest.mock import MagicMock
 
-from sundew import test, test_writer
+from sundew import test, test_writer, utils
 from tests import fixtures
 
 test(test_writer.mock_function_dependencies)(
     setup={fixtures.extend_config_with_dependent_functions},
     kwargs={"fn": fixtures.callee_func, "stack": ExitStack()},
-    returns={
-        "dependent_func": test_writer.DependentFunctionSpy(fixtures.dependent_func)
-    },
+    returns={"dependent_func": utils.FunctionSpy(fixtures.dependent_func)},
 )
 
 test(test_writer.generate_naive_function_import)(
@@ -26,18 +24,12 @@ test(test_writer.generate_function_dependency_test_file)(
     kwargs={
         "fn": test_writer.generate_function_dependency_test_file,
         "mocks": {
-            "generate_naive_function_import": test_writer.DependentFunctionSpy(
+            "generate_naive_function_import": utils.FunctionSpy(
                 test_writer.generate_naive_function_import
             ),
-            "build_test_strings": test_writer.DependentFunctionSpy(
-                test_writer.build_test_strings
-            ),
-            "build_import_string": test_writer.DependentFunctionSpy(
-                test_writer.build_import_string
-            ),
-            "write_tests_to_file": test_writer.DependentFunctionSpy(
-                test_writer.write_tests_to_file
-            ),
+            "build_test_strings": utils.FunctionSpy(test_writer.build_test_strings),
+            "build_import_string": utils.FunctionSpy(test_writer.build_import_string),
+            "write_tests_to_file": utils.FunctionSpy(test_writer.write_tests_to_file),
         },
     },
     returns=None,
